requirements.txt
Django
djangorestframework
django-filter
django-cors-headers
djangorestframework-simplejwt
PyJWT
PyMySQL